Lasagna recipe steps and better page links added
First I finished off the steps section to the lasagna recipe page.
I created an ordered list defining each step, and created an individual
unordered list below each step with the instsructions.
I chose to add in the extra unordered lists to create an indentation
with a bullet point to make the steps more readable.

Afterwords I added a home page and lasagna recipe link to both
the home page and lasagna recipe page, and presented them
in an unordered list to give the website a cleaner appearance.

I may change this later on so that recipe pages only link back
to the home page, any only the home page can link to other recipes.
but I'm not sure how I'd like this  aspect of the website
to look just yet.