Changed format of "Home Page" link to look nicer
To give the home page link a more visually appealing look,
I removed it from the unordered list of web page links on my site.
I then placed it outside of the list and added the
<strong> element to it. I made this change to all pages.

I also made the description for the maple salmon recipe slightly
better by adding "flavors" to the end of the sentence and adding
a period to maintain the format of the site.